pyproject.toml
[project]
name = "unpnet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.3",
    "matplotlib>=3.6,<3.8",
    "numpy>=1.24,<1.27",
    "opencv-contrib-python==4.5.5.64",
    "opencv-python==4.5.5.64",
    "pillow==9.2.0",
    "pip>=25.1.1",
    "torch==2.2.0",
    "torchvision==0.17.0",
    "torchaudio==2.2.0",
    "pyyaml==6.0",
    "scipy==1.9.3",
    "tensorboardx>=2.6",
    "tensorflow==2.10.1",
    "tqdm>=4.67.1",
    "vispy==0.11.0",
    "scikit-learn>=1.7.0",
    "faiss-gpu>=1.7.2",
    "timm>=1.0.17",
    "thop>=0.1.1.post2209072238",
]
